refactor(app): GripperCard
