+++ app/src/organisms/Desktop/Devices/GripperCard/AboutGripperSlideout.tsx
