+++ app/src/organisms/Desktop/Devices/GripperCard/__tests__/AboutGripperSlideout.test.tsx
