+++ app/src/organisms/Desktop/Devices/GripperCard/__tests__/GripperCard.test.tsx
