+++ app/src/organisms/Desktop/Devices/GripperCard/index.tsx
